backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.1.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
stripe==7.7.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.33.6
